Changed component size attribute from Vec2i to Vec2f

+++ include/axl.gl/gfx/UI/Container.hpp
@@ -16,7 +16,7 @@ class AXLGLCXXAPI Container : public axl::gl::gfx::ui::Component
 			axl::gl::gfx::ui::Container* container = 0,
 			Layout* layout = 0,
 			const axl::math::Vec3f& position = axl::math::Vec3f(0.0f,0.0f,0.0f),
-			const axl::math::Vec2i& size = axl::math::Vec2i(0,0),
+			const axl::math::Vec2f& size = axl::math::Vec2f(0,0),
 			const axl::math::Vec4f& padding = axl::math::Vec4f(0.0f,0.0f,0.0f,0.0f));
 		virtual ~Container();
 	public:
