+++ include/axl.gl/gfx/UI/Element.hpp
@@ -5,7 +5,7 @@
 #include "../../camera/Camera3D.hpp"
 #include "Layout.hpp"
 #include <axl.util/ds/List.hpp>
-#include <axl.math/vec/Vec2i.hpp>
+#include <axl.math/vec/Vec2f.hpp>
 #include <axl.math/vec/Vec3f.hpp>
 #include <axl.math/vec/Vec4f.hpp>
 
@@ -32,14 +32,14 @@ class AXLGLCXXAPI Element : public axl::gl::gfx::ui::Component
 			axl::gl::Context* ptr_context = 0,
 			axl::gl::gfx::ui::Container* container = 0,
 			const axl::math::Vec3f& position = axl::math::Vec3f(0.0f,0.0f,0.0f),
-			const axl::math::Vec2i& size = axl::math::Vec2i(0,0),
+			const axl::math::Vec2f& size = axl::math::Vec2f(0,0),
 			const axl::math::Vec4f& padding = axl::math::Vec4f(0.0f,0.0f,0.0f,0.0f),
 			axl::gl::gfx::ui::Layout::Size layout_width = axl::gl::gfx::ui::Layout::WRAP_CONTENT,
 			axl::gl::gfx::ui::Layout::Size layout_height = axl::gl::gfx::ui::Layout::WRAP_CONTENT);
 		virtual ~Element();
 	public:
 		virtual bool isValid() const;
-		virtual axl::math::Vec2i getContentSize() const;
+		virtual axl::math::Vec2f getContentSize() const;
 		axl::gl::gfx::ui::Layout::Size getLayoutWidth() const;
 		axl::gl::gfx::ui::Layout::Size getLayoutHeight() const;
 		void setLayoutWidth(axl::gl::gfx::ui::Layout::Size layout_width);
